+++ lib/shopify_cli/context.rb
@@ -17,6 +17,7 @@ module ShopifyCLI
     GEM_LATEST_URI = URI.parse("https://rubygems.org/api/v1/versions/shopify-cli/latest.json")
     VERSION_CHECK_SECTION = "versioncheck"
     LAST_CHECKED_AT_FIELD = "last_checked_at"
+    LATEST_VERSION_FIELD = "latest_version"
     VERSION_CHECK_INTERVAL = 86400
 
     class << self
@@ -619,11 +620,13 @@ module ShopifyCLI
     #            : nil otherwise
     #
     def new_version
-      if (time_of_last_check + VERSION_CHECK_INTERVAL) < (now = Time.now.to_i)
-        update_time_of_last_check(now)
-        latest_version = retrieve_latest_gem_version
-        latest_version unless latest_version == ShopifyCLI::VERSION
+      if (time_of_last_check + VERSION_CHECK_INTERVAL) < (Time.now.to_i)
+        fork do
+          retrieve_latest_gem_version
+        end
       end
+      latest_version = ShopifyCLI::Config.get(VERSION_CHECK_SECTION, LATEST_VERSION_FIELD, default: ShopifyCLI::VERSION)
+      latest_version unless latest_version == ShopifyCLI::VERSION
     end
 
     # Returns file extension depending on OS
@@ -669,17 +672,15 @@ module ShopifyCLI
     def retrieve_latest_gem_version
       response = Net::HTTP.get_response(GEM_LATEST_URI)
       latest = JSON.parse(response.body)
-      latest["version"]
+      ShopifyCLI::Config.set(VERSION_CHECK_SECTION, LATEST_VERSION_FIELD, latest["version"])
     rescue
       nil
+    ensure
+      ShopifyCLI::Config.set(VERSION_CHECK_SECTION, LAST_CHECKED_AT_FIELD, Time.now.to_i)
     end
 
     def time_of_last_check
       (val = ShopifyCLI::Config.get(VERSION_CHECK_SECTION, LAST_CHECKED_AT_FIELD)) ? val.to_i : 0
-    end
-
-    def update_time_of_last_check(time)
-      ShopifyCLI::Config.set(VERSION_CHECK_SECTION, LAST_CHECKED_AT_FIELD, time)
     end
   end
 end